(Feat): Refactor out metrics code to metrics library
