+++ .vscode/settings.json
@@ -4,5 +4,5 @@
 		"."
 	],
 	"python.testing.unittestEnabled": false,
-	"python.testing.pytestEnabled": true
+	"python.testing.pytestEnabled": true,
 }
