+++ package.json
@@ -56,6 +56,7 @@
     "jest-environment-jsdom": "^29.7.0",
     "jest-junit": "^16.0.0",
     "jsdom": "~22.1.0",
+    "metrics": "^0.1.21",
     "nx": "^20.3.2",
     "prettier": "^2.6.2",
     "react-refresh": "^0.10.0",
