--- packages/common/.python-version
@@ -1,1 +1,0 @@
-3.12.8
