--- packages/common/common/hello.py
@@ -1,6 +1,0 @@
-"""Sample Hello World application."""
-
-
-def hello():
-    """Return a friendly greeting."""
-    return "Hello models"
