+++ packages/common/pyproject.toml
@@ -23,6 +23,7 @@ group = { }
 
   [tool.poetry.dependencies]
   python = ">=3.12,<3.13"
+  pydantic = "^2.10.6"
 
 [build-system]
 requires = [ "poetry-core" ]
